Simplify Eureka Instancer's initial cache priming
Rather than first fetching a Eureka application's instances
synchronously and then scheduling subsequent background updates,
instead schedule the updates, wait on the first update to arrive, an
then proceed.

Revise the tests accordingly to no longer rely on a set of instances
known to the (fake) server as distinct from any applications. Since
applications are an intensional view of instances, like the real
Eureka server we derive the presence of applications from the
application names mentioned by instances. That is, without at least
one instance declaring that it's a member of a given application, no
such application exists.